move focusIssue into shared utils and update the import across different files
