+++ src/applications/appeals/10182/tests/utils/focus.unit.spec.js
@@ -4,7 +4,7 @@ import { render, waitFor } from '@testing-library/react';
 
 import { $ } from 'platform/forms-system/src/js/utilities/ui';
 
-import { focusIssue } from '../../utils/focus';
+import { focusIssue } from '../../../shared/utils/focus';
 import { LAST_ISSUE } from '../../constants';
 
 describe('focusIssue', () => {
@@ -78,6 +78,7 @@ describe('focusIssue', () => {
   it.skip('should focus on second remove button', async () => {
     window.sessionStorage.setItem(LAST_ISSUE, '1,remove-cancel');
     const { container } = await renderPage();
+
     await focusIssue(0, container);
     await waitFor(() => {
       const target = $('#issue-1 .remove-issue', container);
